--- src/main/java/com/zerobase/nsbackend/auth/config/WebClientConfig.java
@@ -1,44 +1,0 @@
-package com.zerobase.nsbackend.auth.config;
-
-import io.netty.channel.ChannelOption;
-import io.netty.handler.timeout.ReadTimeoutHandler;
-import io.netty.handler.timeout.WriteTimeoutHandler;
-import java.time.Duration;
-import java.util.function.Function;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.reactive.ClientHttpConnector;
-import org.springframework.http.client.reactive.ReactorClientHttpConnector;
-import org.springframework.http.client.reactive.ReactorResourceFactory;
-import org.springframework.web.reactive.function.client.WebClient;
-import reactor.netty.http.client.HttpClient;
-
-@Configuration
-public class WebClientConfig {
-    @Bean
-    public ReactorResourceFactory resourceFactory() {
-        ReactorResourceFactory factory = new ReactorResourceFactory();
-        factory.setUseGlobalResources(false);
-        return factory;
-    }
-
-    @Bean
-    public WebClient webClient() {
-
-        Function<HttpClient, HttpClient> mapper = client -> {
-
-            return HttpClient.create()
-                .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, 1000)
-                .doOnConnected(connection -> {
-                    connection.addHandlerLast(new ReadTimeoutHandler(10))
-                        .addHandlerLast(new WriteTimeoutHandler(10));
-                })
-                .responseTimeout(Duration.ofSeconds(1));
-        };
-
-        ClientHttpConnector connector =
-            new ReactorClientHttpConnector(resourceFactory(), mapper);
-
-        return WebClient.builder().clientConnector(connector).build();
-    }
-}
