+++ src/main/java/com/zerobase/nsbackend/auth/utils/KakaoUserInfo.java
@@ -3,26 +3,33 @@ package com.zerobase.nsbackend.auth.utils;
 import com.zerobase.nsbackend.auth.dto.KakaoUserInfoResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.WebClient;
-import reactor.core.publisher.Flux;
+import org.springframework.web.client.RestTemplate;
 
 @RequiredArgsConstructor
 @Component
 public class KakaoUserInfo {
-    private final WebClient webClient;
     @Value("${spring.security.oauth2.client.provider.kakao.user-info-uri}")
     private String USER_INFO_URI;
 
     public KakaoUserInfoResponse getUserInfo(String token) {
         String uri = USER_INFO_URI;
+        RestTemplate restTemplate = new RestTemplate();
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.set("Authorization", "Bearer " + token);
+        HttpEntity request = new HttpEntity(httpHeaders);
 
-        Flux<KakaoUserInfoResponse> response = webClient.get()
-            .uri(uri)
-            .header("Authorization", "Bearer " + token)
-            .retrieve()
-            .bodyToFlux(KakaoUserInfoResponse.class);
+        ResponseEntity<KakaoUserInfoResponse> responseEntity = restTemplate.exchange(
+            uri,
+            HttpMethod.GET,
+            request,
+            KakaoUserInfoResponse.class
+        );
 
-        return response.blockFirst();
+        return responseEntity.getBody();
     }
 }
